requirements.txt
pvlib==0.10.2
dataclasses-json==0.6.3
python-dotenv==1.0.0